First Attempt At Properly Generating Normals
I want to make it so that triangles have a clockwise winding order: their direction is determined by the order with which their points are given. This in turn affects the direction of their normal.

This solution will most likely need to be revisited, in particular because it's generating the normal every time it's accessed...