Some Progress towards Initializing from Mesh working
Normals are showing up now - the code for getting the points of a triangle made a bad assumption about which of the triangle's edges' points made up the three unique vertices of the triangle. I now just explicitly store the points too.

Unfortunately, in some cases the order is ending up wrong now, and the normals are flipped.