It Works!
It turns out in the previous commit I was just a few silly mistakes away from it working. I made it optional to 'collapse co-located points' (might rename this) - basically whether points in the same location will be treated as the same point or not. For cubic shapes this was all I needed. For the sphere primitive, it was already good.

Next, I have to make the 'IsConnected' method work well with morphs where co-located points are not collapsed, since these will technically be treated as disconnected points. I can just make these a special type of edge.