Live Mesh Manipulation!
Expanded notion of 'contiguous faces' to include faces which share a single point and are conormal.

Changed '.Position' to '.LocalPosition' to better convey that this is the mesh's vertex position in its local space.

Spent a lot of quite unsatisfying time looking at the retriangulate method

Added == and != operators for point, and Equals/GetHashCode too, which work on the point IDs, I'm not so sure about those ones.

And, best of all, started making an editor class for MorphTest (which at some point I'll turn into MorphFilter or something) which has handles for mesh manipulation with the cursor, it's cool!