FINALLY got proper connectivity working, perimeters working on concave faces
Took me long enough, but connectivity dictionary now also includes edges where the points aren't actually on same edge, but one of the points is in the same position.

Having this more holistic view of connectivity was what I needed to make the perimeter-generating function work for concave faces, so now it does. Nice!