Retriangulation! (And removing stuff!)
First and foremost, finally adding a functioning retriangulation method. Even if it is just the kind of bad one I used in the previous prototype, it's working nicely enough.

It's fully set up too - it replaces the previous face with the new retriangulated one. This meant finally getting around to adding removal methods for taking faces, triangles, edges, points, out of the morph. They're not fully tested yet.

Next I'd like to make it so moving points will regenerate faces, but in a smart way so that it doesn't do it for every face, only the ones affected.